updated version to 1.8.3 changed packager, so heltzel doesn't get blamed for anything i did wrong changed requires from php-mysql to php, since tiki supports databases other than mysql included standard hack to prevent catastrophe if $RPM_BUILD_ROOT is / removed chown, since %defattr handles it modified %preun rms, since they run in / changed /* to /var/www/html/tiki-1.8 in %files, so tikiwiki doesn't try to own /var, /var/www, and /var/www/html
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@4740 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki/doc/devtools/tikiwiki.spec
@@ -1,7 +1,7 @@
 # Command to build: rpmbuild -ba --target noarch tikiwiki.spec
 
 %define name tikiwiki
-%define version 1.8.RC2
+%define version 1.8.3
 %define release 1
 
 Summary: A PHP-based CMS/Groupware web application with a full Wiki environment
@@ -10,9 +10,9 @@ Version: %{version}
 Release: %{release}
 Copyright: LGPL
 URL: http://www.tikiwiki.org
-Packager: Dennis Heltzel <tiki-rpm@fishroom.net>
+Packager: Peter Popovich <tiki-rpm@popovich.net>
 AutoReqProv: no
-Requires: php-mysql 
+Requires: php 
 Group: Applications/Internet
 Source: http://prdownloads.sourceforge.net/%{name}/%{name}-%{version}.tar.gz
 BuildRoot: %{_tmppath}/%{name}-root
@@ -27,15 +27,12 @@ TikiWiki is an open source CMS/Groupware web application which provides a full W
 %build
 
 %install
-rm -rf $RPM_BUILD_ROOT
+[ "$RPM_BUILD_ROOT" != "/" ] && rm -rf $RPM_BUILD_ROOT
 mkdir -p $RPM_BUILD_ROOT/var/www/html
 cd $RPM_BUILD_ROOT/var/www/html
 tar xvzf $RPM_SOURCE_DIR/%{name}-%{version}.tar.gz
-#tar xvzf $RPM_SOURCE_DIR/%{name}-%{version}%{release}.tar.gz
 mv %{name}-%{version} tiki-1.8
-#mv %{name}-%{version}%{release} tiki-1.8
-# Change ownership and permissions
-chown -R apache:apache *
+# Change file and directory permissions
 cd tiki-1.8
 find . -name "*.php" -exec chmod 644 {} \;
 find . -name "*.sql" -exec chmod 644 {} \;
@@ -47,17 +44,17 @@ find . -name "CVS" -type d -print|xargs rm -rf
 find . -name ".cvsignore" -exec rm -f {} \;
 
 %clean
-rm -rf $RPM_BUILD_ROOT
+[ "$RPM_BUILD_ROOT" != "/" ] && rm -rf $RPM_BUILD_ROOT
 
 %preun
 # Remove unneeded files
-rm -rf templates_c/*
-rm -f modules/cache/*.cache
+rm -rf /var/www/html/templates_c/*
+rm -f /var/www/html/modules/cache/*.cache
 
 %files
 %defattr(-,apache,apache)
 %config /var/www/html/tiki-1.8/db/tiki-db.php
 #%doc /var/www/html/tiki-1.8/README
-/*
+/var/www/html/tiki-1.8
 
 %changelog
